Create DeleteNodeFromLinkedList.java
There is a singly-linked list head and we want to delete a node node in it.
We are given the node to be deleted node. We have not got access to the first node of head.
All the values of the linked list are unique, and it is guaranteed that the given node node is not the last node in the linked list.